Add way to auto install optionnal depndency with pip
